+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Allow just one flow segment for calculateMissingParametersSCSUH endpoint
 - travel_time_sheet_flow equation to match newest spreadsheet
 - traveltimetc example request
+- weightedCurveNumber function now processes published Curve Number data
 
 ### Deprecated